nmos: change the flags values to use &20 for CR mode as it is ' '
In the code that handles CR mode, i.e. whether CR is visible as an
inverse 'M', the value used to test the option flags is subsequently
sent to OSWRCH.  There is no harm if this an ASCII space but it causes
problems if it is another control character so this commit makes the
flag value &20, i.e. ASCII space by swapping with the Insert/Overwrite
flag.